+++ src/structures/ModalSubmitField.js
@@ -1,5 +1,10 @@
 const BaseMessageComponent = require('./BaseMessageComponent');
 
+/**
+ * Represents a Field of a Modal Submit Interaction.
+ * @extends BaseMessageComponent
+*/
+
 class ModalSubmitField extends BaseMessageComponent {
   constructor(data = {}) {
     super({ type: 'TEXT_INPUT' });