+++ src/structures/ShowModal.js
@@ -1,23 +1,29 @@
 const { Client, Interaction } = require('discord.js');
 const { InteractionResponseTypes } = require('../util/Constants');
 const Modal = require('./Modal');
+const { Error } = require('./errors');
 
 /**
  * Shows the Modal to the Interaction User.
  * @param {Modal} Modal Modal.
  * @param {Object} Options Client and Interaction.
+ * @example
+ * showModal(Modal, {
+ *   client: Client, // Client to show the Modal through the Discord API.
+ *   interaction: Interaction // Interaction to show the Modal with the Interaction ID & Token.
+ * });
  * @returns {Modal} Modal.
 */
 
 async function showModal(modal, options){
 
-    if (!modal) throw new Error('[discord-modals] MODAL_REQUIRED: No modal was provided.');
-    if (!options) throw new Error('[discord-modals] OPTIONS_REQUIRED: No options were provided.');
-    if (!options.client) throw new Error('[discord-modals] CLIENT_REQUIRED: No client was provided on the options.');
-    if (!options.interaction) throw new Error('[discord-modals] INTERACTION_REQUIRED: No interaction was provided on the options.');
-    if (!(modal instanceof Modal)) throw new Error('[discord-modals] INVALID_MODAL: The provided modal is invalid.');
-    if (!(options.client instanceof Client)) throw new Error('[discord-modals] INVALID_CLIENT: The provided client is invalid.');
-    if (!(options.interaction instanceof Interaction)) throw new Error('[discord-modals] INVALID_INTERACTION: The provided interaction is invalid.');
+    if (!modal) throw new Error('MODAL_REQUIRED');
+    if (!options) throw new Error('OPTIONS_REQUIRED');
+    if (!options.client) throw new Error('CLIENT_REQUIRED');
+    if (!options.interaction) throw new Error('INTERACTION_REQUIRED');
+    if (!(modal instanceof Modal)) throw new Error('INVALID_MODAL');
+    if (!(options.client instanceof Client)) throw new Error('INVALID_CLIENT');
+    if (!(options.interaction instanceof Interaction)) throw new Error('INVALID_INTERACTION');
 
     const _modal = modal instanceof Modal ? modal : new Modal(modal, options.client);