+++ CHANGELOG.md
@@ -2,7 +2,14 @@
 
 Hello! This is the Change log section of Discord-Modals. Here you can see what things have been changed in the different versions to be informed. This will be updated every time there are new versions. I hope you like this package :)
 
-## v1.1.8 (Stable)
+## v1.1.9 (Stable)
+
+- Add examples to classes. More comprehension.
+- Improvements to documentation and fix errors.
+- More organization in errors.
+- Update readme.
+
+## v1.1.8
 
 - Fix compatibility issues. Now the package throws an error if you have a different version than discord.js v13 to avoid confusions.
 - Update readme.